Sessions History Fix it

+++ hiitfire/Models/SessionsHistoryModel.swift
@@ -6,10 +6,10 @@ struct SessionsHistory: Identifiable, Codable {
     let id: UUID
     var title: String
     var duration: Int
-    var caloriesBourned: Double
+    var caloriesBourned: Int
     var date: Date
     
-    init(id: UUID, title: String, duration: Int, caloriesBourned: Double, date: Date) {
+    init(id: UUID, title: String, duration: Int, caloriesBourned: Int, date: Date) {
         self.id = id
         self.title = title
         self.duration = duration
